pyproject.toml
[project]
name = "joka"
version = "0.1.1"
description = "Joka - MySQL Migration Tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncmy>=0.2.10",
    "build>=1.3.0",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "greenlet>=3.3.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "typer>=0.20.0",
]

[project.scripts]
joka = "joka.__main__:app"

[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["joka*"]
